move memory manager to segment-spi
